backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
motor = "^3.5.1"
beanie = "^1.26.0"
requests = "^2.32.3"
sentry-sdk = "^2.13.0"
pyjwt = "^2.9.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-core = "2.20.1"
pydantic-settings = "^2.4.0"
emails = "^0.6"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.24.0"
mypy = "^1.8.0"
ruff = "^0.2.2"
pre-commit = "^3.6.2"
types-passlib = "^1.7.7.20240106"
coverage = "^7.4.3"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
